Splitting out ControllerStatus Tests into another test class
